requirements.txt
altair==4.1.0
astor==0.8.1
attrs==19.3.0
backcall==0.1.0
base58==2.0.0
beautifulsoup4==4.9.0
bleach==3.1.4
blinker==1.4
bokeh==2.0.1
boto3==1.12.39
chardet==3.0.4
click==7.1.1
decorator==4.4.2
defusedxml==0.6.0
docutils==0.15.2
emoji==0.5.4
entrypoints==0.3
enum-compat==0.0.3
future==0.18.2
gTTS==2.1.1
gTTS-token==1.1.3
idna==2.9
importlib-metadata==1.6.0
ipykernel==5.2.0
ipython==7.13.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.16.0
Jinja2==2.11.2
jmespath==0.9.5
joblib==0.14.1
jsonschema==3.2.0
jupyter-client==6.1.3
jupyter-core==4.6.3
MarkupSafe==1.1.1
mistune==0.8.4
nbconvert==5.6.1
nbformat==5.0.5
nltk==3.5
notebook==6.0.3
numpy==1.18.2
packaging==20.3
pandas==1.0.3
pandocfilters==1.4.2
parso==0.7.0
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.1.1
pkg-resources==0.0.0
prometheus-client==0.7.1
prompt-toolkit==3.0.5
protobuf==3.11.3
ptyprocess==0.6.0
pydeck==0.3.0
Pygments==2.6.1
pyparsing==2.4.7
pyrsistent==0.16.0
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3.1
pyzmq==19.0.0
regex==2020.4.4
requests==2.23.0
s3transfer==0.3.3
Send2Trash==1.5.0
six==1.14.0
soupsieve==2.0
streamlit==0.57.3
terminado==0.8.3
testpath==0.4.4
textblob==0.15.3
toml==0.10.0
toolz==0.10.0
tornado==5.1.1
tqdm==4.45.0
traitlets==4.3.3
typing-extensions==3.7.4.2
tzlocal==2.0.0
urllib3==1.25.8
validators==0.14.3
watchdog==0.10.2
wcwidth==0.1.9
webencodings==0.5.1
widgetsnbextension==3.5.1
zipp==3.1.0
